+++ CHANGELOG.md
@@ -5,6 +5,10 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+
+### Deprecated
+- Deprecated `resource-label` attibute on `manifold-data-product-logo`. Use `manifold-data-resource-logo` component instead.
+
 ### Added
 - Added a SSO data button and the resource wrapper for ssoing into a resource's product dashboard.
 - Added a new CTA slot in the product card for displaying unique cta content.
@@ -12,6 +16,11 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Fixed
 - Fixed the provision button requiring the label to be set, preventing or automatic label generation from working.
 - Fixed name for `manifold-service-view` to be `manifold-service-card-view` to match documentation.
+- Added missing support for theme variable `--manifold-tag-free-text-color`.
+
+### Changed
+- Added graphqlFetch to `manifold-connection`.
+- Converted `manifold-data-product-logo` to use GraphQL.
 
 ## [0.4.3]
 ### Fixed